requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
comm==0.1.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipykernel==6.18.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.1.0
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
slack==0.0.2
slack-bolt==1.15.5
slack-sdk==3.19.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.13
wcwidth==0.2.5
yarl==1.8.1
